requirements.txt
# YouTube & Data
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.2

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
greenlet>=3.0.0

# NLP / AI
google-genai>=1.2.0
pydantic>=2.0.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0

# Resilience
tenacity>=8.2.0

# CLI
typer>=0.9.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0
